initalize 'mypatches' git repo
The purpose of this repository is to collect patches left here and there
to avoid losing them. Even if the best approach would be to upstream all
of them, this is not always feasible. This is no good reason to leave the
work to its deadfull destiny... Hope this repo would give these patches
a stronger hope of surviving